Refactor wastewater queries to separate file
